requirements.txt
alembic==1.17.2
apispec==6.9.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-smorest==0.46.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.1.1
packaging==25.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
webargs==8.7.1
Werkzeug==3.1.4
pytest==8.3.3
pytest-flask==1.3.0
faker==30.8.1
bcrypt
flake8==7.3.0
flask-swagger-ui==4.11.1
